Use type and id as method argument name to correspond to the API request parameters.
We previously used id, knowing it shadowed id() builtin, but renamed
type parameters, and for pcap id used report_id.  In rethinking this,
just use id and type and if we need to use the builtin in a method we
can save a copy in the constructor.

NOTE: not backward compatible.